refs(alert_rules): Stop writing to unused incident fields. Modify data model to prepare for removing unused fields
This refactors the `Incident` model to make `aggregation` and `query` nullable, removes the `groups`
relations and disables the constraints on `IncidentGroup`. We don't use any of these fields any
more, so may as well remove.

We also make `alert_rule` not null, since we will always have an alert rule available now. The data
migration will remove any incidents without alert rules - these are all test rows and are fine to
remove.

All tables involved are small, and so it's fine to do these operations as part of the migration.